changes to motors, joystick, pwm, main
git-svn-id: https://localhost/igvc@704 b8439097-6f30-0410-869d-b57f63bf1bb9

+++ 2010/software/trunk/motors/osmc/arduino/serial_comm_helper.hpp
@@ -3,8 +3,8 @@
 
 #define TIMEOUT_LENGTH_MILLIS 1000
 
-bool serialReadBytesTimeout(int len, byte * msg);
+bool serialReadBytesTimeout(byte len, byte * msg);
 
-void serialPrintBytes(void *data, int numBytes);
+void serialPrintBytes(void *data, size_t numBytes);
 
 #endif
